testCheckpointWithFunction
